+++ Babylon/Assets/1Game/1Pack/ScriptPackage/DunGen/Code/Async/YieldReason.cs
@@ -1,9 +1,0 @@
-namespace DunGen.Async
-{
-	public enum YieldReason
-	{
-		WorkBudget,
-		RoomPlaced,
-		BetweenSteps,
-	}
-}