+++ Babylon/Assets/1Game/1Pack/ScriptPackage/DunGen/Code/Constants.cs
@@ -1,9 +1,0 @@
-namespace DunGen
-{
-	public static class Constants
-	{
-		public const string DefaultDungeonRootName = "Dungeon Root";
-		public const string DefaultDungeonName = "Dungeon";
-	}
-}
-