+++ Babylon/Assets/1Game/1Pack/ScriptPackage/DunGen/Code/Generation/ITileInjector.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace DunGen.Generation
-{
-	public interface ITileInjector
-	{
-		void InjectTiles(RandomStream randomStream, ref List<InjectedTile> tilesToInject);
-	}
-}