+++ Babylon/Assets/1Game/1Pack/ScriptPackage/DunGen/Code/TileBounds/ITileBoundsCalculator.cs
@@ -1,9 +1,0 @@
-using UnityEngine;
-
-namespace DunGen.TileBounds
-{
-	public interface ITileBoundsCalculator
-	{
-		Bounds CalculateLocalBounds(GameObject tileRoot);
-	}
-}